refactor backend routes and controllers
