+++ backend/middlewares/authMiddleware.js
@@ -1,0 +1,16 @@
+import jwt from "jsonwebtoken";
+
+const JWT_SECRET = "zebify_super_secreto";
+
+export default function authenticateToken(req, res, next) {
+  const authHeader = req.headers["authorization"];
+  const token = authHeader && authHeader.split(" ")[1];
+
+  if (!token) return res.sendStatus(401);
+
+  jwt.verify(token, JWT_SECRET, (err, user) => {
+    if (err) return res.sendStatus(403);
+    req.user = user;
+    next();
+  });
+}
