+++ backend/routes/authRoutes.js
@@ -1,0 +1,6 @@
+import { signup, login } from "../controllers/authController.js";
+
+export default function registerAuthRoutes(app) {
+  app.post("/signup", signup);
+  app.post("/login", login);
+}
